Review feedback
* Move Yup schema to separate file
* Remove most tests (and skip the rest)
* Move paginated methods to Manager
