+++ src/services/support/index.ts
@@ -1,9 +1,7 @@
 export {
   getTickets,
-  getTicketsPage,
   getTicket,
   getTicketReplies,
-  getTicketRepliesPage,
   createSupportTicket,
   createReply,
   uploadAttachment,
